requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-keycloak==0.1.1
ecdsa==0.15
gunicorn==20.0.4
idna==2.9
josepy==1.3.0
mozilla-django-oidc==1.2.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-jose==3.1.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
